Rework of the setup.sh script
More fine-grained implementation of the smbus pkg installation
via APT. Script will attempt to install 'smbus' for both Py2 and
Py3, then it will check what was installed to instruct users about
which python version to use.

Reorganized part of the code to improve readability and make it
easier to edit in the future, in case something changes. Examples
include moving CONFIG variables to the top and the addition of a
message() function.