Tiny dashboard Author: [@jdarias](https://github.com/jdarias).
This is a script that shows a famous quote, a currency conversion pair of your choice and the weather of a city. It also shows the last three characters from your ip address, the date in DDMM format and the hour in HH:MM format

<p align="center">
  <img src="imgs/demo_iphour.gif">
</p>

The script takes info from the following apis and shows it in the 16\*2 display.

* [theysaidso.com](https://theysaidso.com): Free public API that provides famous quotes from well known people. It has a public endpoint that doesn't require an API key.

* [exchangerate-api.com](https://exchangerate-api.com) / [free.currencyconverterapi.com](https://free.currencyconverterapi.com): There are a lot of currency apis but these ones offer free currency exchange info. Both are used, one as main, the other as backup.

* [openweathermap.org](https://openweathermap.org): Provides Weather info, forecasts, etc.

In order to use the script you need to get API key tokens for both exchange rate services and the weather api. Open the script in a text editor and put your tokens in the USER VARIABLES section.

Also set a currency exchange pair. For currency support and the currency codes you need to use, see [Here for free.currencyconverter.com](https://free.currencyconverterapi.com) (click the "List of all currencies" link) and [Here for exchangerate-api.com](https://www.exchangerate-api.com/docs/supported-currencies)

A city/country string is also needed to show weather info for such city. Search for your city on openweathermap.org and take note of the name,country and put it in the script. London,gb is given as an example.

I, [@jdarias](https://github.com/jdarias) am not affiliated to, or responsible of the information services used in this script.

Code Explanation:
The script uses four APIs, two for currency conversion, one for famous
quotes, and another one for weather. Since each api has different update
restrictions, I need to create four different threads to request the
information according to each api limitation. ie. weather api info can
be requested every 5 minutes, but exchange rate api updates once a day.
So each thread will request info from the api, then pause for a certain
time, until the next update.
The info obtained is passed to four global variables in the main thread.
This thread starts the four child api threads one after another. When
one thread passes the info to the corresponding global variable, the
next thread is started, until all four threads are running.
Then the main thread starts a loop that will display the info from the
global variables on the lower line of the display, together with the
hour, date and ip from the device on the upper line of the display.